Quick fix - java.lang.IncompatibleClassChangeError
Our com.urbancode.air.ExitCodeException class in the project air/plugins is throwing a java.lang.IncompatibleClassChangeError on java 7.  com.urbancode.air.ExitCodeException extends Exception and the binaries used in this plug-in where compiled using Java 6. In java 7 the Exception class got an extra constructor. Groovy checks things such as the amount on constructors and runtime hence the error.

We need to rewirte the exception in Java. There is no good work around except to not use it.